Dev v0.9

+++ src/main/java/com/demo/riskproject/dto/request/TaskRequest.java
@@ -18,5 +18,5 @@ public class TaskRequest {
 
     @Min(1)
     @Max(5)
-    private Byte point;
+    private Integer point;
 }
